# Parallel output in tpchgen-cli (Nx faster, where N is number of cores)