@alamb 2025-03-25
BTW with the other improvements on main, when merging it takes less than 5 seconds on my laptop to make TPCH SF 10 🚀 

```shell
time target/release/tpchgen-cli -s 10  --output-dir=/tmp/tpchdbgen-rs

real	0m4.895s
user	0m36.096s
sys	0m2.910s
```